index.php erstellt und mit list.php verknüpft

+++ clear.php
@@ -2,7 +2,16 @@
 
 require_once "inc/database_functions.inc.php";
 
-deleteAllItems();
+// Nur POST-Anfragen erlauben
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
-header("Location: list.php");
+    /**
+     * Alle Datensätze löschen
+     */
+    deleteAllItems();
+}
+
+// Zurück zur Liste
+
+header('Location: list.php');
 exit;
